Fixed bugs on Chart

+++ Nitride/Source/Tools/Time.cs
@@ -23,7 +23,7 @@ namespace Nitride
 
         public static double ToEpoch(this DateTime time) => (time.ToUtc() - Epoch).TotalSeconds;
 
-        public static DateTime FromEpoch(long epoch) => Epoch.AddSeconds(epoch);
+        public static DateTime FromEpoch(double epoch) => Epoch.AddSeconds(epoch);
 
         public static (bool valid, DateTime time) ToDateTime(this string input, string format = "")
         {
